+++ server/handlers/custom-route-handler.js
@@ -14,6 +14,7 @@ const staticPageTemplate = ejs.compile(staticPageTemplateStr);
 const { CustomPath } = require("../impl/api-client-impl");
 const { createBasicStore } = require("./create-store");
 const { addCacheHeadersToResult } = require("./cdn-caching");
+const { addStaticPageMimeType } = require("./mime-type-handler");
 
 function renderStaticPageContent(store, content) {
   const renderedContent = staticPageTemplate({ store, content });
@@ -66,6 +67,7 @@ exports.customRouteHandler = function customRouteHandler(
     logError,
     seo,
     domainSlug,
+    cdnProvider = null
   }
 ) {
   const url = urlLib.parse(req.url, true);
@@ -84,7 +86,7 @@ exports.customRouteHandler = function customRouteHandler(
             "Defaulting the status-code to 302 with destination-path as home-page"
           );
         }
-        addCacheHeadersToResult(res, page.cacheKeys(config["publisher-id"]));
+        addCacheHeadersToResult({ res: res, cacheKeys: page.cacheKeys(config["publisher-id"]), cdnProvider: cdnProvider, config: config });
 
         let destination = page["destination-path"] || "/";
 
@@ -97,7 +99,8 @@ exports.customRouteHandler = function customRouteHandler(
       }
 
       if (page.type === "static-page") {
-        addCacheHeadersToResult(res, page.cacheKeys(config["publisher-id"]));
+        addCacheHeadersToResult({ res: res, cacheKeys: page.cacheKeys(config["publisher-id"]), cdnProvider: cdnProvider, config: config });
+        addStaticPageMimeType({ res, page });
 
         if (page.metadata.header || page.metadata.footer) {
           return loadData("custom-static-page", {}, config, client, {